+++ pydfs_lineup_optimizer/lineup_printer.py
@@ -1,4 +1,4 @@
-from typing import cast, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 
 if TYPE_CHECKING:
@@ -11,7 +11,8 @@ class BaseLineupPrinter:
 
 
 class LineupPrinter(BaseLineupPrinter):
-    OUTPUT_FORMAT = '{0:>2}. {1:<5} {2:<30}{3:<6}{4:<15}{5:<9}{6:<8}{7:<10}\n'
+    OUTPUT_FORMAT = '{index:>2}. {lineup_position:<5} {name:<30}{positions:<6}{team:<15}{game:<9}' \
+                    '{fppg:<15}{salary:<10}\n'
 
     def _print_game_info(self, player: 'LineupPlayer') -> str:
         game_info = player.game_info
@@ -50,45 +51,9 @@ class LineupPrinter(BaseLineupPrinter):
         return res
 
 
-class DropLowestLineupPrinter(LineupPrinter):
-    @staticmethod
-    def _get_lowest_fppg_player(lineup: 'Lineup') -> 'LineupPlayer':
-        return cast('LineupPlayer', sorted(lineup, key=lambda p: p.fppg)[0])
-
-    def _print_player(self, index: int, player: 'LineupPlayer', is_dropped: bool = False) -> str:
-        return self.OUTPUT_FORMAT.format(
-            index,
-            player.lineup_position,
-            '%s%s' % (player.full_name, '(DROPPED)' if is_dropped else ''),
-            '/'.join(player.positions),
-            player.team,
-            self._print_game_info(player),
-            round(player.fppg, 3),
-            str(player.salary) + '$',
-        )
-
-    def _print_footer(self, lineup: 'Lineup') -> str:
-        footer = super(DropLowestLineupPrinter, self)._print_footer(lineup)
-        footer += 'Fantasy Points Without Dropped Player %.2f' % \
-                  (lineup.fantasy_points_projection - self._get_lowest_fppg_player(lineup).fppg)
-        return footer
-
-    def print_lineup(self, lineup):
-        res = ''
-        lowest_fppg_player = self._get_lowest_fppg_player(lineup)
-        for index, player in enumerate(lineup.players, start=1):
-            res += self._print_player(index, player, is_dropped=player == lowest_fppg_player)
-        res += '\n'
-        res += self._print_footer(lineup)
-        return res
-
-
 class IndividualSportLineupPrinter(LineupPrinter):
-    def _print_player(self, index, player):
-        return '{0:>2}. {1:<5} {2:<30}{3:<8}{4:<10}\n'.format(
-            index,
-            player.lineup_position,
-            player.full_name,
-            round(player.fppg, 3),
-            str(player.salary) + '$',
-        )
+    OUTPUT_FORMAT = '{index:>2}. {lineup_position:<5} {name:<30}{fppg:<15}{salary:<10}\n'
+
+
+class DraftKingTiersLineupPrinter(LineupPrinter):
+    OUTPUT_FORMAT = '{index:>2}. {lineup_position:<5} {name:<30}{fppg:<15}\n'