+++ pydfs_lineup_optimizer/player.py
@@ -4,7 +4,6 @@ from pytz import timezone
 from typing import List, Optional
 from pydfs_lineup_optimizer.utils import process_percents
 from pydfs_lineup_optimizer.tz import get_timezone
-from pydfs_lineup_optimizer.constants import PlayerRank
 
 
 GameInfo = namedtuple('GameInfo', ['home_team', 'away_team', 'starts_at', 'game_started'])
@@ -19,7 +18,6 @@ class Player:
                  team: str,
                  salary: float,
                  fppg: float,
-                 rank: PlayerRank = PlayerRank.REGULAR,
                  is_injured: bool = False,
                  actual: Optional[float] = None,
                  max_exposure: Optional[float] = None,
@@ -30,6 +28,9 @@ class Player:
                  min_deviation: Optional[float] = None,
                  max_deviation: Optional[float] = None,
                  is_confirmed_starter: Optional[bool] = None,
+                 fppg_floor: Optional[float] = None,
+                 fppg_ceil: Optional[float] = None,
+                 original_positions: Optional[List[str]] = None,
                  ):
         self.id = player_id
         self.first_name = first_name
@@ -55,6 +56,9 @@ class Player:
         self.max_deviation = max_deviation
         self.projected_ownership = projected_ownership
         self.is_confirmed_starter = is_confirmed_starter
+        self.fppg_floor = fppg_floor
+        self.fppg_ceil = fppg_ceil
+        self._original_positions = original_positions
 
     def __repr__(self):
         return '%s %s (%s)' % (self.full_name, '/'.join(self.positions), self.team)
@@ -129,13 +133,16 @@ class Player:
                 return True
         return False
 
+    @property
+    def original_positions(self) -> List[str]:
+        return self._original_positions or self.positions
+
 
 class LineupPlayer:
-    __slots__ = ['_player', 'lineup_position']
-
-    def __init__(self, player: Player, lineup_position: str):
+    def __init__(self, player: Player, lineup_position: str, used_fppg: Optional[float] = None):
         self._player = player
         self.lineup_position = lineup_position
+        self.used_fppg = used_fppg
 
     def __getattr__(self, attr_name):
         return getattr(self._player, attr_name)